+++ je_auto_control/utils/action_signing/__init__.py
@@ -3,14 +3,17 @@ from je_auto_control.utils.action_signing.cipher import (
     decrypt_action_file, encrypt_action_file,
 )
 from je_auto_control.utils.action_signing.signer import (
-    VerifyResult, require_signed_actions, sign_action_file, verify_action_file,
+    VerifyResult, read_signed_action_bytes, require_signed_actions,
+    sign_action_file, signing_required, verify_action_file,
 )
 
 __all__ = [
     "VerifyResult",
     "decrypt_action_file",
     "encrypt_action_file",
+    "read_signed_action_bytes",
     "require_signed_actions",
     "sign_action_file",
+    "signing_required",
     "verify_action_file",
 ]